test case  x86

+++ test/fastmath.jl
@@ -275,7 +275,7 @@ end
 end
 # issue #53857
 @testset "fast_pow" begin
-    n = 2^52
+    n = Int64(2)^52
     @test @fastmath (1 + 1 / n) ^ n ≈ ℯ
     @test @fastmath (1 + 1 / n) ^ 4503599627370496 ≈ ℯ
 end
